+++ src/index.ts
@@ -32,6 +32,7 @@ export {
   type ClientConversationStore,
   type NavigationCheckpoint,
   SessionStorageStore,
+  type ThreadMeta,
 } from "./core/conversation_store.js";
 export {
   type AgentFactory,
@@ -39,6 +40,7 @@ export {
   type HttpAgentOptions,
 } from "./core/create_http_agent.js";
 export { defineAgUiChat } from "./core/define_ag_ui_chat.js";
+export { RemoteConversationStore } from "./core/remote_conversation_store.js";
 export {
   type FlashOptions,
   focusWithFlash,